EQ-251 Moved the values endpoints to the /api/attains/ url.
