+++ app/server/app/routes/attains.js
@@ -377,6 +377,145 @@ function executeQueryCountOnly(profile, req, res) {
   }
 }
 
+/**
+ * Retrieves the domain values for a single table column.
+ * @param {express.Request} req
+ * @param {express.Response} res
+ */
+function executeValuesQuery(req, res) {
+  const profile = tableConfig[req.params.profile];
+  if (!profile) {
+    return res
+      .status(404)
+      .json({ message: 'The requested profile does not exist' });
+  }
+
+  const { additionalColumns, ...params } = getQueryParamsValues(req);
+
+  const columnAliases = [
+    req.params.column,
+    ...(Array.isArray(additionalColumns) ? additionalColumns : []),
+  ];
+
+  const columns = [];
+  for (const alias of columnAliases) {
+    const column = profile.columns.find((col) => col.alias === alias);
+    if (!column) {
+      return res.status(404).json({
+        message: `The column ${alias} does not exist on the selected profile`,
+      });
+    }
+    columns.push(column);
+  }
+
+  queryColumnValues(profile, columns, params, req.activeSchema)
+    .then((values) => res.status(200).json(values))
+    .catch((error) => {
+      log.error(
+        `Failed to get values for the "${req.params.column}" column from the "${profile.tableName}" table: ${error}`,
+      );
+      res.status(500).send('Error! ' + error);
+    });
+}
+
+/**
+ * Gets the query parameters from the request.
+ * @param {express.Request} req
+ * @returns request query parameters
+ */
+function getQueryParamsValues(req) {
+  return Object.entries(req.method === 'POST' ? req.body : req.query).reduce(
+    (current, [key, value]) => {
+      if (key in current) {
+        return { ...current, [key]: value };
+      } else if (req.method === 'GET') {
+        return { ...current, filters: { ...current.filters, [key]: value } };
+      } else return current;
+    },
+    {
+      text: '',
+      direction: null,
+      filters: {},
+      limit: null,
+      additionalColumns: [],
+    },
+  );
+}
+
+/**
+ * Craft the database query for distinct column values
+ * @param {Object} profile definition of the profile being queried
+ * @param {Array<Object>} columns definitions of columns to return, where the first is the primary column
+ * @param {Object} params parameters to apply to the query
+ * @param {string} schema the currently active database schema
+ */
+async function queryColumnValues(profile, columns, params, schema) {
+  const primaryColumn = columns[0];
+
+  // get columns for where clause
+  const columnsForFilter = [];
+  profile.columns.forEach((col) => {
+    if (params.filters.hasOwnProperty(col.alias)) {
+      columnsForFilter.push(col.name);
+    }
+  });
+
+  // search through tableconfig.materializedViews to see if the column
+  // we need is in here
+  const materializedView = profile.materializedViews.find((mv) => {
+    for (const col of columnsForFilter.concat(columns.map((c) => c.name))) {
+      if (!mv.columns.find((mvCol) => mvCol.name === col)) return;
+    }
+    return mv;
+  });
+
+  // query table directly if a suitable materialized view was not found
+  const query = knex
+    .withSchema(schema)
+    .from(materializedView ? materializedView.name : profile.tableName)
+    .column(
+      columns.reduce(
+        (current, col) => ({ ...current, [col.alias]: col.name }),
+        {},
+      ),
+    )
+    .whereNotNull(primaryColumn.name)
+    .distinctOn(primaryColumn.name)
+    .orderBy(primaryColumn.name, params.direction ?? 'asc')
+    .select();
+
+  // build where clause of the query
+  profile.columns.forEach((col) => {
+    appendToWhere(query, col.name, params.filters[col.alias]);
+  });
+
+  if (params.text) {
+    query.andWhere((q) => {
+      columns.forEach((col, i) => {
+        if (col.type === 'numeric' || col.type === 'timestamptz') {
+          i === 0
+            ? q.whereRaw('CAST(?? as TEXT) ILIKE ?', [
+                col.name,
+                `%${params.text}%`,
+              ])
+            : q.orWhereRaw('CAST(?? as TEXT) ILIKE ?', [
+                col.name,
+                `%${params.text}%`,
+              ]);
+        } else {
+          i === 0
+            ? q.whereILike(col.name, `%${params.text}%`)
+            : q.orWhereILike(col.name, `%${params.text}%`);
+        }
+      });
+    });
+  }
+
+  if (params.limit) query.limit(params.limit);
+
+  return await query;
+}
+
 export default function (app, basePath) {
   const router = express.Router();
 
@@ -406,6 +545,15 @@ export default function (app, basePath) {
         executeQueryCountOnly(profile, req, res);
       },
     );
+
+    // get column domain values
+    router.get('/:profile/values/:column', function (req, res) {
+      executeValuesQuery(req, res);
+    });
+
+    router.post('/:profile/values/:column', function (req, res) {
+      executeValuesQuery(req, res);
+    });
   });
 
   app.use(`${basePath}api/attains`, router);
